Refactor: Extracted BookCellView to separate file
