+++ Audiobook Player/Library/BookCellView.swift
@@ -1,0 +1,16 @@
+//
+//  BookCellView.swift
+//  Audiobook Player
+//
+//  Created by Florian Pichler on 12.04.18.
+//  Copyright © 2018 Tortuga Power. All rights reserved.
+//
+
+import UIKit
+
+class BookCellView: UITableViewCell {
+    @IBOutlet weak var artworkImageView: UIImageView!
+    @IBOutlet weak var titleLabel: UILabel!
+    @IBOutlet weak var authorLabel: UILabel!
+    @IBOutlet weak var completionLabel: UILabel!
+}
